+++ lib/types/DataSources.ts
@@ -15,7 +15,7 @@ export interface GroupCacheConfiguration extends CommonCacheConfiguration {
 
 export type CacheEntry<LoadedValue> = {
   key: string
-  value: LoadedValue
+  value: LoadedValue | null
 }
 
 /**
@@ -91,17 +91,22 @@ export interface GroupCache<LoadedValue> extends GroupWriteCache<LoadedValue> {
 
 /**
  * Data source interface for retrieving values.
- *
- * Return value semantics:
- * - Return the actual value when found
- * - Return `null` to indicate "value was resolved but is empty" - this WILL be cached
- * - Return `undefined` to indicate "value was not resolved" - this will NOT be cached,
- *   and the next data source in the sequence will be queried
  */
 export interface DataSource<LoadedValue, LoadParams = string, LoadManyParams = LoadParams extends string ? undefined : LoadParams> {
+  /**
+   * - Return the actual value when found
+   * - Return `null` to indicate "value was resolved but is empty" - this WILL be cached
+   * - Return `undefined` to indicate "value was not resolved" - this will NOT be cached,
+   *   and the next data source in the sequence will be queried
+   */
   get: (loadParams: LoadParams) => Promise<LoadedValue | undefined | null>
 
-  // note that we cannot combine keys and loadParams here, because we may be asked to only fetch a subset of originally requested keys, as the others might be cached already
+  /**
+   * Omit keys that did not resolve - returning a `null` for them is not supported here.
+   *
+   * Note that we cannot combine keys and loadParams here, because we may be asked to only fetch a
+   * subset of originally requested keys, as the others might be cached already.
+   */
   getMany: (keys: string[], loadParams?: LoadManyParams) => Promise<LoadedValue[]>
 
   name: string
@@ -109,15 +114,16 @@ export interface DataSource<LoadedValue, LoadParams = string, LoadManyParams = L
 
 /**
  * Group data source interface for retrieving values within groups.
- *
- * Return value semantics:
- * - Return the actual value when found
- * - Return `null` to indicate "value was resolved but is empty" - this WILL be cached
- * - Return `undefined` to indicate "value was not resolved" - this will NOT be cached,
- *   and the next data source in the sequence will be queried
  */
 export interface GroupDataSource<LoadedValue, LoadParams = string, LoadManyParams = LoadParams extends string ? undefined : LoadParams> {
+  /**
+   * - Return the actual value when found
+   * - Return `null` to indicate "value was resolved but is empty" - this WILL be cached
+   * - Return `undefined` to indicate "value was not resolved" - this will NOT be cached,
+   *   and the next data source in the sequence will be queried
+   */
   getFromGroup: (loadParams: LoadParams, group: string) => Promise<LoadedValue | undefined | null>
+  /** Omit keys that did not resolve - returning a `null` for them is not supported here. */
   getManyFromGroup: (keys: string[], group: string, loadParams?: LoadManyParams) => Promise<LoadedValue[]>
 
   name: string